predictive-health-guardian/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
streamlit==1.39.0
crewai==0.51.1
langchain-huggingface==0.1.2
transformers==4.46.3
torch==2.5.1
faker==30.2.0
scikit-learn==1.5.2
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
pydantic==2.9.2
numpy==1.26.4
